# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Camila SaaS - API para emprendedores locales
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores/me:
    get:
      operationId: getMyStore
      tags: [stores]
      summary: Get current store details
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateMyStore
      tags: [stores]
      summary: Update store profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStoreRequest"
      responses:
        "200":
          description: Store updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/me/users:
    get:
      operationId: getStoreUsers
      tags: [stores]
      summary: Get users of current store
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: addStoreUser
      tags: [stores]
      summary: Add a user to the store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stores/me/users/{userId}:
    delete:
      operationId: removeStoreUser
      tags: [stores]
      summary: Remove a user from the store
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    UpdateStoreRequest:
      type: object
      properties:
        businessName:
          type: string
        address:
          type: string
        district:
          type: string
        phone:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        primaryColor:
          type: string
        description:
          type: string
        whatsapp:
          type: string
        socialInstagram:
          type: string
        socialFacebook:
          type: string
...
    AddUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [store_admin, store_staff, cashier]
      required:
        - name
        - email
        - password
        - role
...
    User:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [superadmin, store_admin, store_staff, cashier]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - name
        - email
        - role
        - isActive
        - createdAt
    Store:
      type: object
      properties:
        id:
          type: string
        businessName:
          type: string
        businessType:
          type: string
          enum: [clothing, restaurant, bakery, fair_booth, general_catalog]
        documentType:
          type: string
          enum: [DNI, RUC10, RUC20]
        documentNumber:
          type: string
        ownerName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        district:
          type: string
        logoUrl:
          type: string
        bannerUrl:
          type: string
        primaryColor:
          type: string
        description:
          type: string
        whatsapp:
          type: string
        socialInstagram:
          type: string
        socialFacebook:
          type: string
        isActive:
          type: boolean
        license:
          $ref: "#/components/schemas/License"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - businessName
        - businessType
        - documentType
        - documentNumber
        - ownerName
        - phone
        - email
        - district
        - isActive
        - createdAt
    License:
      type: object
      properties:
        id:
          type: string
        storeId:
          type: string
        status:
          type: string
          enum: [trial, active, expired, suspended]
        startsAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        notes:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - storeId
        - status
        - updatedAt
...