# lib/api-zod/src/generated/types/addUserRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Camila SaaS - API para emprendedores locales
 * OpenAPI spec version: 0.1.0
 */
import type { AddUserRequestRole } from "./addUserRequestRole";

export interface AddUserRequest {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  role: AddUserRequestRole;
}